price-reporter: manager: Refactor into module
