+++ core/src/main.rs
@@ -31,8 +31,8 @@ use job_types::proof_manager::new_proof_manager_queue;
 use job_types::task_driver::new_task_driver_queue;
 use job_types::{event_manager::new_event_manager_queue, gossip_server::new_gossip_server_queue};
 use network_manager::{worker::NetworkManager, worker::NetworkManagerConfig};
-use price_reporter::worker::ExchangeConnectionsConfig;
-use price_reporter::{manager::PriceReporter, worker::PriceReporterConfig};
+use price_reporter::worker::PriceReporterConfig;
+use price_reporter::worker::{ExchangeConnectionsConfig, PriceReporter};
 use proof_manager::{proof_manager::ProofManager, worker::ProofManagerConfig};
 use state::{create_global_state, tui::StateTuiApp};
 use system_bus::SystemBus;
